minor: Renamed module `arithmetic` to `arithmetic-expression`
